Commit a few small fixes for bugs found by findbugs: 1) Capture stream read return values in an int, as a char cannot be negative and thus cannot handle a -1 error return value. 2) When calling String.replaceAll(), capture the output as it does not alter the original String. 3) Cast int value to a long before shifting left by 32 bits.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@823206 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/pack200/src/main/java/org/apache/harmony/unpack200/SegmentHeader.java
@@ -283,7 +283,7 @@ public class SegmentHeader {
     private void parseArchiveFileCounts(InputStream in) throws IOException,
             Pack200Exception {
         if (options.hasArchiveFileCounts()) {
-            setArchiveSize(decodeScalar("archive_size_hi", in, Codec.UNSIGNED5) << 32
+            setArchiveSize((long)decodeScalar("archive_size_hi", in, Codec.UNSIGNED5) << 32
                     | decodeScalar("archive_size_lo", in, Codec.UNSIGNED5));
             archiveSizeOffset = in.available();
             setSegmentsRemaining(decodeScalar("archive_next_count", in,
